Make the UIWebView extension a signal instead
